+++ RadixSort.h
@@ -10,5 +10,5 @@ using namespace chrono;
 
 double RadixSortRuntime(int* arr, int n);
 
-void Com_RadixSort(int* arr, int n, int& cnt_com);
+void Com_RadixSort(int* arr, int n, long long& cnt_com);
