+++ flashSort.h
@@ -6,9 +6,9 @@
 using namespace std;
 using namespace chrono;
 
-int minElement(int *arr, int n, int &count_comparison);
-int maxElement(int *arr, int n, int &count_comparison);
-void flashSortComparison(int *arr, int n, int &count_comparison);
+int minElement(int *arr, int n, long long &count_comparison);
+int maxElement(int *arr, int n, long long &count_comparison);
+void flashSortComparison(int *arr, int n, long long &count_comparison);
 
 void flashSort(int *arr, int n);
 double flashSortRuntime(int *arr, int n);
